+++ android/src/main/java/com/dexterous/flutterlocalnotifications/BitmapSource.java
@@ -1,0 +1,6 @@
+package com.dexterous.flutterlocalnotifications;
+
+public enum BitmapSource {
+    Drawable,
+    FilePath
+}